+++ ALSystemUtilities/ALSystemUtilities/ALNetwork/ALNetwork.m
@@ -356,8 +356,8 @@
     NSUInteger BA = ~NM | IP;
     
     // Make a string for the address
-    NSString *BroadcastAddress = [NSString stringWithFormat:@"%d.%d.%d.%d", (BA & 0xFF000000) >> 24,
-                                  (BA & 0x00FF0000) >> 16, (BA & 0x0000FF00) >> 8, BA & 0x000000FF];
+    NSString *BroadcastAddress = [NSString stringWithFormat:@"%ld.%ld.%ld.%ld", (long)((BA & 0xFF000000) >> 24),
+                                  (long)((BA & 0x00FF0000) >> 16), (long)((BA & 0x0000FF00) >> 8), (long)(BA & 0x000000FF)];
     
     // Check to make sure the string is valid
     if (BroadcastAddress == nil || BroadcastAddress.length <= 0) {
